[PATCH] ppc64: Make pci_alloc_consistent() conform to API docs
Documentation/DMA-mapping.txt says that pci_alloc_consistent() needs to
return a mapping that is aligned by the closest larger order of two as the
allocation.

We're currently breaking this with our iommu code.  To fix this, add
align_order arguments to the relevant functions and pass it down.
Specifying align_order of 0 gives same behaviour as previous.

Signed-off-by: Olof Johansson <olof@austin.ibm.com>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 419e7a11T360jOXIeor2KEvsvnPzYQ