[PATCH] ppc64: Fix default command line
The ppc64 kernel can be built with a default command line (CONFIG_CMDLINE)
for cases where none is provided by the firmware.

However, some OF implementation always pass a "bootargs" property that only
contains the "0" terminating byte of a C string which caused us to think
there was a command line, and not use the built-in one. This patch fixes it.

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 41a14ce28bj4zOQWr2CyKte7L13-vA