[PATCH] SELinux: map Unix seqpacket sockets to appropriate security class
This patch for SELinux fixes a bug in the mapping of socket types to
security classes and ensures that Unix seqpacket sockets are mapped to an
appropriate security class.  The Unix stream security class is re-used in
this case as it has the same permission checking applied as for seqpacket.

Signed-off-by:  Stephen Smalley <sds@epoch.ncsc.mil>
Signed-off-by:  James Morris <jmorris@redhat.com>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 41a21712YKiGKiFBpse2vEvO-RlQvQ