[PATCH] ppc64: Make early processor spinup based on physical ids
This changes the early CPU spinup code to be based on physical CPU ID
instead of logical.  This will make it possible to kexec off of a
different cpu than 0, for example after it's been hot-unplugged.

The booted cpu will still be mapped as logical cpu 0, since there's various
stuff in the early boot that assumes logical boot cpuid is 0.

Also, it expands the kexec boot param structure to allow the booted physical
cpuid to be passed in.  This includes bumping the version number to 2 for
backwards compat.

Signed-off-by: Olof Johansson <olof@austin.ibm.com>
Signed-off-by: Anton Blanchard <anton@samba.org>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 41a671d9_IULR1rghOCSA6wfsM7-DA