[PATCH] cciss: Off-by-one error causing oops in CCISS_GETLUNIFO ioctl
This patch fixes an an "off-by-one" error found in the CCISS_GETLUNIFO
ioctl in the cciss driver.  It is cycling through the part table of the
gendisk structure which is a zero-based array, not a one-based array.  This
often causes an oops when referencing the out-of-bounds element.

Signed-off by: Andrew Patterson <andrew.patterson@hp.com>

Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 41ab9716MMs-_X01WX6p1L3srnYmcw