[PATCH] md: fix careless bug in raid10
raid10 performs resync by reading all blocks, comparing to look for The
requires re-tasking a bio that was used to read, so that it can be re-used
for write.  So a number of fields need to be reset.

The bi_bdev field was not being reset.  However when using on a partition
of a device, it gets changed by generic_make_request.

This patch simply resets bi_bdev to the correct value.

Signed-off-by: Neil Brown <neilb@cse.unsw.edu.au>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 41ae6b01sihj2ciz2o6jhl-4iRxfKA