[PATCH] Documentation for IDE and CDROM ioctls
I've written two document files, and packaged them as a patch.

Two things caught my eye while I was writing this up:

The header comments for CDROMREADRAW, CDROMREADMODE1, and CDROMREADMODE2
disagree with the actual source code.  The header comments imply that a
cdrom_read structure is used to pass data, but the source code actually
reads a cdrom_msf structure and then overwrites it with raw data.  I'm not
sure if there's a bug in the header comments, in the driver source, or I
misread something.

The CDROM_LOCKDOOR ioctl seems to lock/unlock all doors on all drives,
because it uses a global variable.

Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 41ae6bb2OXpYPasXJpTtgAm6FLmzhA