Merge bk://bk.arm.linux.org.uk/linux-2.6-rmk into ppc970.osdl.org:/home/torvalds/v2.6/linux
2004/12/03 22:32:50+00:00 rmk
[ARM] Add per_cpu data area to linker script.

2004/12/03 22:22:04+00:00 rmk
[ARM] Ensure user ops pass 64-bit constants in even,odd registers.

Always pass values to get_user and put_user in an even numbered
register, and optionally the next odd numbered register.  This
ensures that we are compatible with compiler enhancements.

2004/12/03 16:38:42+00:00 ben-linux
[ARM PATCH] 2290/1: S3C2410 - timer usec fixes

Patch from Ben Dooks

A simple divide is not enough for the s3c2410 default
timer code, due to most clocks not being a nice multiple
of usecs.

This update changes the calculation of the usec to use
multipliers and shifts to effect a fast divide by a non-integer
number.

Original patch by Dimitry Andric, updated by Ben Dooks

Signed-off-by: Dimitry Andric

Signed-off-by: Ben Dooks
Signed-off-by: Russell King

BKrev: 41b21dfaocFc59xJcJIo3wNaD1pn2g