[PATCH] restore BLKRRPART semantics
In 2.6.8 the code for the BLKRRPART ioctl was changed
to return EIO when no partitions were found, such as
on an empty disk. This breaks some partitioning programs
and is also confusing: "Input/Output error" while in fact
nothing was wrong with this brand new all blank disk.

This restores old behaviour.

BKrev: 41b5f369U4m7qWO7xkzbqVANQTTW1A