Merge bk://gkernel.bkbits.net/libata-2.6 into ppc970.osdl.org:/home/torvalds/v2.6/linux
2004/12/07 19:23:59-05:00 dougg
[PATCH] off-by-1 libata-scsi INQUIRY VPD pages 0x80 and 0x83

I have some code (in sginfo) that requests the first 4 bytes
of SCSI INQUIRY VPD pages to get their length then asks for
that exact length in a follow up request to fetch the payload.
Just like I saw with 36 byte standard INQUIRYs (no fixed)
I get a buffer  full or zeroes.

BTW SCSI standards dictate that in situations where the allocation
length (in the cdb) is less than what is needed that what can be
sent shall be sent (i.e. truncated and without any error indication
or modification to the part of the response returned).
In other words it is up the the application client to take remedial
action.

Changelog:
   - fix off-by-1 allocation length issue with SCSI
     INQUIRY VPD pages 0x80 and 0x83

Signed-off-by: Jeff Garzik <jgarzik@pobox.com>

2004/12/07 18:54:34-05:00 jgarzik
[libata] only DMA map data for DMA commands (fix >=4GB bug)

libata made the assumption that (for PIO commands in this case)
it could modify DMA memory at the kernel-virtual address, after
mapping this.  This is incorrect, and fails on e.g. platforms that
copy DMA memory back and forth (swiotlb on Intel EM64T and IA64).

Remove this assumption by ensuring that we only call the DMA mapping
routines if we really are going to use DMA for data xfer.

Also:  remove a bogus WARN_ON() in ata_sg_init_one() which caused
bug reports (but no problems).

BKrev: 41b66ad3ERG9o3gvIP0yfxCyYe2h_Q