Merge bk://kernel.bkbits.net/davem/net-2.6 into ppc970.osdl.org:/home/torvalds/v2.6/linux
2004/12/08 22:55:27-08:00 wensong
[IPVS] add a sysctl variable to expire quiescent template

The patch is from Horms <horms@vergenet.net>

Signed-off-by: Horms <horms@verge.net.au>
Signed-off-by: David S. Miller <davem@davemloft.net>

2004/12/08 22:37:57-08:00 davem
Merge http://linux-mh.bkbits.net/bluetooth-2.6
into nuts.davemloft.net:/disk1/BK/net-2.6

2004/12/08 22:00:32-08:00 davem
[IPV6]: Do not use udp_poll for RAW sockets.

Signed-off-by: David S. Miller <davem@davemloft.net>

2004/12/08 21:37:13-08:00 solar
[TCP]: Missing KERN_INFO in remotely triggerable printk.

Signed-off-by: David S. Miller <davem@davemloft.net>

2004/12/08 20:35:33-08:00 shemminger
[PKT_SCHED]: netem forgets to restart device after inserting packets

The version of netem in 2.6.10 moves packets from the delayed queue
to the qdisc in a timer interrupt. But it forgot to force the device to
pick them up.

Signed-off-by: Stephen Hemminger <shemminger@osdl.org>
Signed-off-by: David S. Miller <davem@davemloft.net>

2004/12/08 20:33:03-08:00 kaber
[PKT_SCHED]: Fix oops in ipt action error path.

This patch fixes an oops when the ipt action is used with a
non-existant iptables target. It tries to log
t->u.kernel.target->name, u.kernel.target is part of a union
and as long as the target wasn't successfully loaded contains
the name of the target, using it as a pointer results in a
crash.

Signed-off-by: Patrick McHardy <kaber@trash.net>
Signed-off-by: David S. Miller <davem@davemloft.net>

2004/12/08 20:31:35-08:00 jt
[IRDA]: Use kill_urb() in irda-usb

* change comment about Sigmatel now that there is a driver
* convert to new module_param
* places where urb is unlinked synchronously, use usb_kill_urb
  because that is now a runtime warning.

Original patch from Stephen Hemminger

Signed-off-by: Stephen Hemminger <shemminger@osdl.org>
Signed-off-by: Jean Tourrilhes <jt@hpl.hp.com>
Signed-off-by: David S. Miller <davem@davemloft.net>

2004/12/08 20:29:41-08:00 jt
[IRDA]: Use kill_urb() in stir4200

USB changed behaviour of unlink_urb so that it gives warning and backtrace when
being used synchronously. The correct current behaviour is to us kill_urb
in that case.

Original patch from Stephen Hemminger

Signed-off-by: Stephen Hemminger <shemminger@osdl.org>
Signed-off-by: Jean Tourrilhes <jt@hpl.hp.com>
Signed-off-by: David S. Miller <davem@davemloft.net>

2004/12/08 20:28:20-08:00 jt
[IRDA]: Try to fix the worst abuse of the pci init code in via_ircc

Bugfix suggested by Arkadiusz Miskiewicz.

Signed-off-by: Jean Tourrilhes <jt@hpl.hp.com>
Signed-off-by: David S. Miller <davem@davemloft.net>

2004/12/06 10:48:06+01:00 marcel
[Bluetooth] Use module parameter for ISOC alternate setting

The module parameter for the ISOC transfers now chooses the value
of the alternate setting.

Signed-off-by: Marcel Holtmann <marcel@holtmann.org>

2004/12/06 10:39:53+01:00 marcel
[Bluetooth] Track the class of device value

The class of device is only available from an inquiry response or
from the connection request. In the case of an incoming connection
the value can be taken from the request and stored in the hci_conn
structure. For an outgoing connection the value from the inquiry
cache is used or it is set to zero.

Signed-off-by: Marcel Holtmann <marcel@holtmann.org>

2004/12/06 09:18:04+01:00 marcel
[Bluetooth] Use separate inquiry data structure

The inquiry results can return different fields and so create one
data structure that has place for every of these. Missing fields
are filled with default values.

Signed-off-by: Marcel Holtmann <marcel@holtmann.org>

BKrev: 41b8e703ot0fEHYp5eV_sl1V3pgTJg