[PATCH] back out CPU clock additions to posix-timers
This patch reverts the additions of an ABI supporting thread and process
CPU clocks in the posix-timers code.  This returns us to 2.6.9's condition,
there is no support for any new clockid_t values for process CPU clocks.

This also fixes the return value for clock_nanosleep when unsupported (I
think this is used only by sgi-timer at the moment).  The POSIX-specified
code for valid clocks that don't support the sleep operation is ENOTSUP.
On most architectures the kernel doesn't define ENOTSUP and this name is
defined in userland the same as the kernel's EOPNOTSUPP.

Signed-off-by: Roland McGrath <roland@redhat.com>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 41c8b709MPvaQyj3A-432SlGjHDTMA