x86 ptrace: remove long stale (and bitrotted) test for PT_DTRACE without PT_PTRACED.
Long ago, the "D" in PT_DTRACE meant "Delayed", and it was used
as a flag to mark that we had ptrace'd the process but no longer
did so. That hasn't been true in a while now, and the flag should
probably be renamed, but in the meantime the test for PT_PTRACED
being cleared had been corrupted into something totally nonsensical.

Pointed out by Andi Kleen.

BKrev: 41d89d296TeAroR0G8SPCJJAsLnguA